skeleton for Network
